Trait keyremap::KeyEventHandler

pub(crate) trait KeyEventHandler: Debug + Display {
    // Required methods
    fn handle_event(
        &mut self,
        pv: &mut ProcView<'_>
    ) -> Result<(KeyAction, HandlerEvent)>;
    fn reset(&mut self);
    fn get_state(&self) -> HandlerState;
}

Receives keyboard input events and matches them to a set of internal conditions, such as a specific key or group of keys being pressed.

structs of this trait make up the main configuration of the application.

Required Methods

fn handle_event( &mut self, pv: &mut ProcView<'_> ) -> Result<(KeyAction, HandlerEvent)>

fn reset(&mut self)

fn get_state(&self) -> HandlerState